Fixed a bug in App.js
Hi @bjrshussain , I'm really fascinated with this project and your explanation. You're amazing!
However, I've found a little bug in a player which hasn't appeared in your YouTube video.
The problem is when you trying to skip the song to the next, it stops instead of turning the next one.
I've added the value currentSong to array of dependencies useEffect, so know it works correctly.